OM-44947: XL TP: Register ITransport handlers before putting probe into probe store.
Problem
=======
There were cases where the ProbeStore would keep a closed ITransport indefinitely.

Diagnosis
=========
The ProbeStore::registerNewProbe method makes a retrying remote call to consul while holding the inner data lock. In the environment where the bug was observed, this remote call was failing for 2 hours (!!!) while Consul was down.

In RemoteMediationServer, we don't register the onClose() and onMessage() handlers until AFTER the ProbeStore::registerNewProbe method returns. If the transport gets closed in that time, the probe store will never know about it!

Immediate Fix
=============
The immediate fix is to register the handlers as soon as we get the registerTransport call. We should do this anyway - the longer we wait the longer the window of opportunity is for a closed connection to go unnoticed. This is what I'm adding in this change.

Additional Fixes
================
We shouldn't be doing the retrying call to consul while holding the lock - all probe-related operations end up being blocked! In order to handle this properly we need to reduce the locking, while still keeping the RemoteProbeStore thread-safe. For example, after the KV operation is complete we need to check if there have been any requests to remove the probe. I'm deferring this (to OM 45232) to fix the P1 issue.

Testing
=======
- Ran existing unit tests. I don't have an environment where I can reproduce this issue.

Reviewed By: karthikt
https://rbcommons.com/s/VMTurbo/r/30583/