OM-44742: XL Settings: Convert "GetEntitySettings" RPC call to return a stream of responses.
Issue
=====
In large topologies, automation and workflow settings were not being properly applied in the Action Orchestrator.

Background
==========
The **GetEntitySettings** RPC call was being called from the Action Orchestrator on all entities involved in actions. In the RBC topology that amounted to most entities in the topology, and the resulting **GetEntitySettingsResponse** was large enough to violate the maximum gRPC message size.

Solution
========
The fix is to stream the entity settings from the group component to the client. I changed the signature of the **GetEntitySettings** call to return the stream, and set a default "chunk" size of 20.

Testing
=======
- Build passed.
- Tested with RBC topology. Verified that the error (about exceeding max gRPC message size) is gone.

Reviewed By: garyzeng
https://rbcommons.com/s/VMTurbo/r/30813/