OM-45602: XL API: Allow stats requests to operate when repository or plan are down.
Issue
=====
If repository (or plan orchestrator) are down, we are unable to process stat requests - even for things that shouldn't require repo/plan access!

Solution
========
Allow execution to proceed - if the Repo is down the RepositoryApi class will return an empty search result (and log an error). If the plan orchestrator is down the code to check if something is a plan will return an empty optional (assume it's not a plan - in most cases it's not!).

Testing
=======
- Verified that stats can be retrieved when repo and plan are down.

Reviewed By: marcoberlot
https://rbcommons.com/s/VMTurbo/r/30862/