OM-44818 : Reduced the number of repository calls for the breadcrumb api. RB:30658
The UI entity view results in 48 queries to arangodb(because of the
duplicate api calls + the AQL calls we are making in the backend to
return results for these api call). For entities lower in the supply
chain(such as storage, datacenter, etc), the supply chain has a lot of
connected entities. As we are loading all the entity DTOs, it's causing memory
pressure on arangodb. The changes in this commit cuts down the queries to AQL
by 1/3rd. Many of the Storage entity views which were causing
arangodb OOM earlier, can now be loaded after this fix.