OM-45392 [RB-30918] Fix issue with duplicate probeInfos causing a Topology Processor crash loop after loading diags
Problem
=======
When loading diags (customer topologies), we wipe out and replace the contents of the in-memory ID and probeInfo caches--however, we do not make corresponding changes in the values stored in Consul. Once the values in-memory and in Consul get out of sync, it becomes possible for the store in Consul to accumulate multiple probeInfos of the same probe type. When this occurs, Topology Processor continues to function until the next restart, at which time it fails to start up due to the duplicate probeInfos and it enters a boot loop.

Specifically, these are the steps that lead to duplicate probeInfos and the boot loop problem:

1) Start up the system. The mediation probes register with RemoteProbeStore and get assigned IDs in IdentityProviderImpl.
2) Load diags. The IDs assigned to the probes in IdentityProviderImpl get replaced, and the probe map in the RemoteProbeStore gets replaced too. No impact on Consul.
3) Restart mediation probes container. The mediation probes re-register with IdentityProviderImpl, but now they get the IDs assigned to the probe categories in the diags. The probe infos get written to Consul under the new IDs.
4) Restart topology processor. This should hit this exception.

Solution
========
The solution has two parts.

1. Keep Consul and the in-memory caches in sync with each other when loading diags. This is achieved by making the same changes to Consul that we are making to the caches. Specifically, we wipe the entire folder containing probeIds, as well as the one holding probeInfos. We then replace them with the values stored in the diags.
2. Enforce a specific ordering between restoring the IdentityProvider diags and the Probes diags. Specifically, the IdentityProvider diags need to be restored first. Otherwise, a race condition occurs where if a probe (re)registers after the restore of the probe diags and before the restore of the identityProvider diags, then a probeInfo will be written to Consul with the old ID. This entry will later become a duplicate entry and cause severe problems.