OM-45652: XL TP: Allow probe re-registrations to add Identity Metadata for new entity types.
Intent
======
Adding new entity types to a probe's Identity Metadata shouldn't require a server-side migration.

Background
==========
When an already-registered probe re-registers with the TP we compare the existing ProbeInfo with the new ProbeInfo and check several fields for equality. If the fields are not equal we disallow the registration. The list of EntityIdentityMetadata is one of those fields. This means that in order for a probe to be able to discover new entity types we need a server-side migration in the TP (to modify the "existing" probe info in Consul).

This is not ideal - for our own probes, and definitely for third-party probes. Since adding a new entity type doesn't affect the rest of the system

Fix/Implementation
==================
- **ProbeCompatibilityChecker** now has a bunch of compatibility checks. It checks for equality of certain fields, and has special handling for IdentityMetadata:
- If the "new" ProbeInfo contains any entity types that are not in the "existing" probe info, that is allowed.
- Anything else (removing or modifying existing entity types) is not allowed. I created separate issues (45656 and 45657) to support those changes because we need to allow them if we ever want third-party probes!
- **IdentityProviderImpl** is now responsible for applying the compatibility check whenever a **ProbeInfo** maps to an existing **ProbeId**. This is because we need to update the per-probe identity metadata inside **IdentityProviderImpl** only when the new metadata is compatible with the existing one.
- **RemoteProbeStore** no longer applies the **ProbeCompatibilityCheck**, but catches the **IdentityProviderException** from the **IdentityProvider**.

Testing
=======
- Full build.
- Brought up topology processor and mediation. Restarted mediation, verified that it re-registered.

Reviewed By: ron.even
https://rbcommons.com/s/VMTurbo/r/30931/