OM-45809: XL Group: Migrate "group_data" column to "MEDIUMBLOB" (from "BLOB").
Issue
=====
When loading large customer diags we saw that the Group protobuf was too large for a "BLOB."

Background
==========
This is probably from groups that contain more than ~8000 entities (which would put the blob definitively over 64kb). In large environments it's very possible to have such large groups.

Implementation
==============
Migrate to "MEDIUMBLOB" which can store blobs of up to 16MB, which
will be enough for any group size.

Testing
=======
- Built with tests.

Reviewed By: garyzeng
https://rbcommons.com/s/VMTurbo/r/31047