OM-46116 : Cache global supply chain in repository inestead of querying it from arangodb. RB:31385
Currently, the global supply chain is created by combining the results
of (1) Querying Arangodb to get the entity types and count (2) In-memory
structure which stores the provider/consumer relationships between the
entityTypes.

(2) was done as traversing the entire graph in arangodb to create the
relationships was deemed expensive. In this change, we improve (1).
Instead of querying arangodb every time for getting global supply chain
entity info, we store it memory. We create and store these 2 structures
in repository component whenever we recieve a new source realtime
topology. These strucutres are also stored in arangodb. On repository
component startup, they are loaded into repository memory from arangodb.