OM-45588: XL API: Only copy latest snapshot as "current" if the request is looking for stats in a time range.
Issue
=====
Requests for "live"/current stats returned the same commodities twice.

Background
==========
For other widgets (e.g. Workload Improvements) we needed to return the current stats so that they could be cross-referenced with action stats. However, we SHOULDN'T be doing this for requests without a time range, because those usually expect just one snapshot in the results.

Implementation
==============
Check for the existence of a time range in the request. Only copy "latest" snapshot if there is a time range.

Testing
======
- Built API component.
- Brought up XL and verified commodities no longer appear twice.

Reviewed By: psoohoo
https://rbcommons.com/s/VMTurbo/r/31618/