OM-46794: XL History: Only apply entity type condition when querying market stats tables.
Issue
=====
NPE in history component when scoping to an entity.

Background
==========
We were looking for an entity_type field in a non-market table. We shouldn't do this.

Implementation
==============
Only apply the related entity type condition in the history component if the table we're working with contains that field.

Testing Done
===========
- Built history component.
- Successfully brought up and viewed a single entity's stats (which was failing before).

Reviewed By: psoohoo
https://rbcommons.com/s/VMTurbo/r/31618/