OM-46670: XL Repo Client: Early-return on empty oid input.
Issue
=====
At BofA searches over certain tabs (Accounts, Billing Families) were taking a long time and showing no data.

Background
==========
Under the covers we were trying to get an "empty" list of entities from the repository, which was treating it as "get all entities."

Fix
===
Fix the RepositoryClient to early-return on an empty input. I verified that none of the existing uses of RepositoryClient rely on the "empty = all" behaviour.

Testing
=======
- Brought up API component. Accounts and Billing Families load instantly (no results).

Reviewed By: psoohoo
https://rbcommons.com/s/VMTurbo/r/31689/