OM-46750: XL Repo: Use correct topology ID when searching plan DTOs.
Issue
=====
Actions were not appearing in plan results.

Background
==========
The implementation of searchPlanEntityDTOs was still using the live database to sort. However, there is no longer a live database (because live entities are in-memory).

Fix
===
Use the plan database for sorting plan entities. This makes sense, because we should be sorting the plan results using the plan's data!

Testing
=======
- Build with existing tests.
- Verified that actions show up in the Plan UI. Note - some actions are missing details, but that's a separate issue.

Reviewed By: psoohoo
https://rbcommons.com/s/VMTurbo/r/31704/