OM-46963: HistorydbIO - apply query timeout setting to internal connection pool RB:31787
The Problem: The usage and capacity charts in the UI would intermittently display as empty.

The Root Cause: When investigating this issue I observed that this chart is populated from information in the <entity_type>_stats_latest (ie pm_stats_latest, vm_stats_latest, etc) tables in the database. When the problem occurred, the only information in these tables was price index information and there would be no commodity information which is what is used to populate the usage and capacity charts.

Initially we suspected the problem was in the database, but investigating via commands like:
show engine innodb status;

or

show processlist;

indicated there were no queries in progress and therefore the History component was waiting on a connection that would never return. Further investigation revealed that when the problem occurred, it would last for slightly over a day, and then the problem would disappear and the charts would be populated again. It seemed odd that these connections were not timing out until a full day had elapsed which led me to investigate the length of the timeout. The query timeout was supposed to be 15 minutes, but when I connected a debugger I observed that the actual value being used was 24 hours, matching the behavior seen here. Looking at the initialization code for historydbIO, it appears that the timeout on the underlying connection pool can only actually be set at initialization time, and then any further attempts to set it only modify a local variable without adjusting the underlying connection pool.

The Solution:
Have requests to set the historydbIO query timeout set the timeout on the underlying connection pool even after the connection pool has been created.