OM-46999: XL Repository: Always return full TopologyEntityDTOs.
There are several places where we need the "full" TopologyEntityDTOs. Right now we are returning partial entities (it was an attempt to save memory), which is breaking some functionality.

There is a separate bug open for a generic way to specify which part of the entity we need. But for now, to avoid various errors, we should return the full entities if those are requested.

Note: Also setting the EnvironmentType in the "partial" entity. It was an oversight not to set it before.

Testing
=======
- Build, with tests.

Reviewed By: ron.even
https://rbcommons.com/s/VMTurbo/r/31781/