OM-47060: XL SQL: Disable JooqTracingInterceptor
Issue
=====
Somehow, the JooqTracingInterceptor seems to be resetting the dialect in some SQL configurations.

Background
==========
We were getting exceptions when using "ON DUPLICATE KEY UPDATE" in the cost component. The exception said the dialect was "DEFAULT", even though I verified (via logs) that the dialect right before calling the query was "MARIADB". I tried disabling the JooqTracingInterceptor, and was able to see costs. I'm not sure how it caused the issue, but I don't have time to investigate further right now.

Note - the JooqTracingInterceptor is also printing errors because I was expecting it to work differently than it is - I assumed that we would always call "end" before a subsequent "start", and that doesn't seem to be happening. It's probably better to disable it until we can figure out how it's actually working.

Testing
========
Ran AWS probe, and verified that with the fix the price table gets saved and there are entity costs calculated.

Reviewed By: garyzeng
https://rbcommons.com/s/VMTurbo/r/31909/