OM-47239: XL Repo: Close streams when compressing/decompressing realtime entities.
Issue
=====
In 7.16.4 and 7.17 repository was getting restarted on a fairly regular basis due to OOM.

Background
==========
The repository component was regularly restarting, even though the GC logs looked reasonable (especially w.r.t. heap memory usage). Karthik suggested we may be leaking resources in off-heap resources, and looking at the code it turned out we weren't closing resources properly when compressing/decompressing realtime entities.

In 7.16.3 this wasn't as much of a problem, because we weren't de-compressing entities often. In 7.16.4 and 7.17 we request "full" TopologyEntityDTOs in several places, which leads to these failures.

Solution
========
Close all the streams!

Testing
=======
- All existing tests pass.
- I'm not going to test whether this fixes the underlying issue locally. It will quickly become evident as we deploy new builds of 7.17.

Reviewed By: karthikt, garyzeng
https://rbcommons.com/s/VMTurbo/r/31931/