OM-46623: XL TP: Make policy and settings-related stages required. Add retry interval capability to pipeline stages.
Issue
=====
Market was generating 400-500k actions, mostly vStorage resizes.

Background
==========
Shravan and I observed that the 400-500k action plan was being generated when the group component was down. Upon further inspection, it turned out that the policy and settings application stages weren't getting the setting and policy definitions. As a result, the Entity Setting Applicators weren't being run, and we were broadcasting a topology where the vStorage commodities were being set as "resizeable."

The immediate fix is to mark the stages involved in policy and setting application as "required". If, for whatever reason, we can't apply policies or settings, we should stop the broadcast. Sending out a topology that doesn't respect user policies or setting policies will lead to wrong actions, invalidate existing "correct" actions, and cause all sorts of chaos.

A longer-term improvement is to keep a cache of the groups, settings, and policies in the topology processor. However, even with the cache we need to consider whether or not to allow broadcasts if settings are not uploaded to the group component. Other parts of the system rely on these settings to decide what to do with the actions.

Implementation
==============
- Make the policy and settings application stages required.
- Add a retry interval (as part of the PassthroughStage). Be default, we will not retry. A stage can override the default retry behaviour if necessary.

Testing
=======
- Ran existing tests.
- Verified that topology no longer gets broadcast if group component is down.

Reviewed By: emanuele, karthikt
https://rbcommons.com/s/VMTurbo/r/31944/