OM-47431 : Remove Flyway setBaselineOnMigrate flag. RB:32126
Remove the flyway setBaselineOnMigrate flag. The default
value is set to false.

When the baselineOnMigrate flag is set to true, if mariadb
shutsdown/crashes while running the history v1.0 baseline migration
script, restarting the migration when the db comes back, results in
partial creation of vmtdb tables as it skips the re-execution of
baseline script. This would mask any issues and would lead to
correctness problems.

So we set the  baselineOnMigrate flag to false. This way,
the baseline script would fail immediately and we can find the problem
sooner instead of having partially created vmtdb schemas.

The fundamental problem is that mariadb doesn't support transactions
with DDL statements. Due to the limitation of mariadb,
there seems to be no way around this issue other than failing the
migration during startup.