OM-47656: XL Repo: Improving concurrency in MarketTopologyEntitiesListener.
Issue
=====
When re-starting the repository, we sometimes weren't skipping any projected topologies!

Background
==========
When restarting the repository on an appliance I noticed that projected topologies weren't being skipped. It seems that when we read several summaries from the analysis-summary topic it is possible for the summaries to arrive to the MarketTopologyEntitiesListener out of order (if they get scheduled on different threads). We should account for that when setting the "latest known projected topology ID."

Also, we weren't synchronizing access to the latest known ID (and it wasn't marked as volatile) so it was possible for other threads to see an outdated version of it.

Testing
=======
- Restarted the repository in my appliance. Verified that stale projected topologies get skipped.

Reviewed By: psoohoo
https://rbcommons.com/s/VMTurbo/r/32177/