OM-42731: XL AO: Allow sorting actions by recommendation time and savings.
Issue
=====
Failed to sort actions by "creation date" (in the Accepted Action widget's tabular format).

Background
==========
We didn't have handling for the "CREATION_DATE" enum in the PaginationMapper, or in the action orchestrator.
Also, I discovered that we weren't properly looking at the savings-per-hour when using the "SAVINGS" sort order, and we weren't using a stable sort - so actions with the same creation date (or category) could come in arbitrary order.

Implementation
==============
- Added a "StableActionComparator" wrapper around all action comparators. Uses action ID if the internal comparator returns 0.
- Handling "CREATION_DATE" (we already save recommendation time, so this is straightforward).

Testing
=======
- Existing tests pass.
- Unit tests for all new functionality.

Reviewed By: garyzeng
https://rbcommons.com/s/VMTurbo/r/32192/