OM-47740: XL: Upgrade gRPC and Jetty to most recent versions.
Our gRPC version is over a year old, which is a long time in a relatively new project. Upgrading may help with some of the performance issues and/or memory issues we've been seeing in concurrent user tests. The gRPC server now uses a newer Netty version as well (4.1.22 -> 4.1.34).

Testing
=======
- Full build
- Brought up XL, added target, triggered broadcast, navigated UI.

Reviewed By: karthikt
https://rbcommons.com/s/VMTurbo/r/32229/