OM-46890: XL Repo: Entity detail levels for repository queries.
INTENT
======
Reduce the memory and CPU overhead of API/UI operations and action mode calculation on the repository.

BACKGROUND
==========
In Patrick and Jeff's concurrent user testing, we were seeing bad performance and high Repository off-heap memory usage (which led to crashes until we changed the heap percentage). This is due to the the way the Repository stores the in-memory topology:

When we receive the topology, we create a **RepoGraphEntity** for every topology entity. We store the most important properties as properties on the POJO, and we store the full TopologyEntityDTO in a compressed byte array. The idea was that we don't need most properties most of the time, so we shouldn't need to de-compress the full TopologyEntityDTO very often. However, we currently have no mechanism to specify which properties we need, so to handle all use cases we DID end up de-compressing the TopologyEntityDTO every time we called **retrieveTopologyEntities** or **searchTopologyEntityDTOs**! This meant higher CPU usage, higher memory footprint in Repo and API (including off-heap, because GZip uses JNI), and lower response times. Also, the Action Orchestrator used **retrieveTopologyEntities** to retrieve all entities related to all actions every broadcast, which used a huge amount of memory and sometimes caused AO OOMs.

SOLUTION
========
I was initially working on a generic way to specify subsets of protobuf fields, but that effort would take too long to bring to completion right now. Looking at the specific use cases, we could approximate the benefits with four hard-coded "detail levels":
1) Minimal - just the name, type, oid, environment type.
2) API - the commonly required fields in the API component - not including commodities or entity properties.
3) "Action" - Basically 1) plus VMEM and VCPU. We may need to add more fields here later on.
4) Full.

Unfortunately this increases coupling between the components, and adds component-specific semantics to the inter-component communication layer. On the other hand, the savings are huge, and because of the way things are named it's fairly clear how things fit together. Also, because different pieces of the code work with different types, it makes it easy to track which detail levels are used where, and easy to detect when we need a property that isn't supported by a particular detail level.

IMPLEMENTATION
==============
**Protobuf**
- Creating a **PartialEntity** message type, which contains the different detail levels of entities supported in the system.
- **Search** - Removing **searchPlanTopologyEntityDTOs** (because users should use RepositoryService), and **searchTopologyEntityDTOs**. Extending **searchEntities** to add a "return type" parameter to indicate the desired detail level. Also adding a **searchEntitiesStream** - a method intended for internal use, to avoid tripping ourselves up over pagination.
- **Repository** - Extending **retrieveTopologyEntities** to add a "return type" parameter.

**API**
- Removing almost all direct uses of RepositoryServiceStub or SearchServiceStub, and using the RepositoryApi utility instead.
   - The only exception is when we need to pass pagination parameters to the search RPC (when sorting entities by name). Since there was only one place we had to do that, I didn't add the pagination-friendly utilities to RepositoryAPI.
- Reimplementing **RepositoryApi** to provide two request types - a "search" request and a request for specific entities. The "search" request uses the SearchService, and the entity-by-id request uses the RepositoryService.
    - Each request has four "exit" points:
       1) Retrieve "minimal" entities.
       2) Retrieve "api partial" entities - the common case.
       3) Retrieve "full" entities - needed for some aspect mappers.
       4) Retrieve "ServiceEntityApiDTO"s - this includes mapping, as well populating severity, aspects (optional), and target information.
- The majority of the changes is just converting all the places that worked with TopologyEntityDTOs or ServiceEntityApiDTOs to use the appropriate detail level.
    - Note - in a couple of the aspect mappers (VirtualVolume and Port) we end up retrieving the full entity.
- Also, since everything is going through RepositoryApi and ServiceEntityMapper now, it's easy to cache the TargetApiDTOs that we create for the mapping.

**Action Orchestrator**
- Converted the EntitiesAndSettingsSnapshotFactory, ActionModeCalculator, and ActionDescriptionBuilder to use the ActionPartialEntity.

**Repository**
- **PartialEntityConverter** responsible for creating various kinds of **PartialEntity** messages from either a **TopologyEntityDTO** or a **RepoGraphEntity** (this is the optimized realtime case).
- Converted the various RPC services to pass things to the **PartialEntityConverter** before returning.

**Topology Processor**
- All callers continue to request full TopologyEntityDTOs. We can try to optimize this later if necessary.
- Using SearchProtoUtil in some existing search calls.

Testing
=======
- All existing tests pass.
- Additional unit tests
- System testing

Reviewed By: psoohoo
https://rbcommons.com/s/VMTurbo/r/32131/