OM-47264: XL AO: Use more generic explanations when grouping action stats by risk.
Issue
=====
BofA - action stats requests grouped by "risk" fail because the Action Orchestrator response was ~20MB - much higher than the gRPC limit.

Background
==========
When grouping by "risk" we were grouping by the description of the action - that is, the action's explanation. The intent is to group all the actions with specific causes (e.g. vCPU Congestion, Increase RI Coverage) together. However, the full explanation for many actions included entity-specific details (the translation block which had the ID). This meant each action would have its own "group", identified by the explanation, and we would be sending thousands - or tens of thousands - of fairly long strings in the stats response. The response was both huge AND useless (because what's the point of grouping when each group only has 1 entry?)!

Fix
===
The fix is to extend the ExplanationComposer to create short explanations if necessary. The short explanation does not include entity details, specific numbers, or commodity key information (so, for example, instead of "entity 1 can be suspended to improve efficiency" we would say "Current supplier can be suspended to improve efficiency").

Testing
=======
- Extended the ExplanationComposer tests to test the short and full description for each case.
- Testing in BofA environment - no errors when going to On-Prem dashboard.

Reviewed By: psoohoo
https://rbcommons.com/s/VMTurbo/r/32265/