OM-45087: Replaced resource paths in tests to point to copies.
Tests should refer to the copies maven creates automatically under
`target/test-classes` instead of the originals under `src/test/resources`
to follow maven conventions and avoid unexpected results.

Did not replace references in comments that pointed to the places where
manual modifications should be made as those should be done on the
originals, not the copies.