OM-47844: XL API: In plans, requesting entities involved in actions from the projected topology.
Issue
=====
Plan actions not showing up in 7.17.0 and develop.

Background
==========
I added an optimization to request entities from the source topology where possible - because in the realtime context it's much cheaper to get entities from the source topology. However, in plans we don't save the source topology, so we were getting no related entity information!

Fix
===
Modifying the RepositoryApi helper class to internally target the projected topology when used with a non-realtime context.

Future Work
===========
Plan actions still don't show up in develop, because of a separate issue. The action descriptions are now supposed to be calculated in the action orchestrator, but that's only been done for the LiveActionStore. Tariq is working on the plan side under this issue: OM 47108

Testing
=======
- New test.
- Existing tests pass.
- Tested running a plan.

Reviewed By: psoohoo
https://rbcommons.com/s/VMTurbo/r/32408/