OM-44702: XL API: Handle global temp groups in action stats queries.
Issue
=====
When clicking on the global supply chain, the "All Actions" widget was not showing action stats, even though they were available.

Background
==========
We do not keep action stats pre-aggregated for groups. However, global temporary groups represent the entire market, filtered by entity types, and we do keep action stats for that. The fix is therefore to detect when the scope for an action stats query is a global temporary group, and format the Action Orchestrator request accordingly.

Implementation
==============
I extended the UuidMapper and the ApiId classes (since they were already used for action stats queries). ApiId now caches not only whether an ID refers to a group, but also whether the group is a global temp group, and the entity type of the group. Also, UuidMapper now caches ApiIds. This means that the type of a particular OID (or UUID) (entity/group/plan) can be shared between requests, as long as they use the UuidMapper.fromUuid() method instead of working with the input scopes directly. The work to actually convert the existing code to use UuidMapper uniformly will be done separately.

Note
====
The original issue is no longer reproducible in 7.17.0, because the global temp groups get created with the "On Prem" environment type, and do not include the "All Actions" widget. I tested the issue by hard-coding the environment type on the server, but there may not be a way to test on a 7.17 instance.

Testing
=======
- Build, with tests, and new tests. 90% coverage.
- Manually changed the temp group to return "HYBRID" type and verified that the global temp group-scoped All Actios widget has historical data.

Reviewed By: psoohoo
https://rbcommons.com/s/VMTurbo/r/32425/