OM-36862: XL API: Set "activeEntitiesCount" in GroupApiDTO.
Issue
=====
The "active" entity count in GroupApiDTO is always the same as the regular entity count.

Fix
===
When mapping a Group to a GroupApiDTO, do a search for "active" entities in the group. We only return the count to minimize traffic on the network. There is no great way to further optimize this for a large number of groups, because we have to run each search independently anyway.

Testing
=======
- Full build, with extra unit tests.
   - I don't know of a place in the UI where we use the active entities count, so I didn't do integration testing.

Reviewed By: marcoberlot
https://rbcommons.com/s/VMTurbo/r/32467/