OM-46605, RB-32520: Target scopes are not supported in the XL stats API
Background
The REST API supports a call POST /stats/{target-id} which is intended to execute the given query over the set of SE's discovered by the fix.

Problem
This case, where the ID indicates a target, was not handled at all in the API Component, StatsService.

Solution
Implement a utility class to fetch the SE's for a target if called with the target ID. The resulting list of SE's is handled in the same way that an expanded Group is handled.