OM-47907: XL Repo: Treat search starting OIDs as starting filter.
The expectation is that the entity OIDs in the SearchEntitiesRequest will limit the initial set of entities considered for the search. However, the TopologyGraphSearchRpcService implementation was applying the OID filter to the results of the search.

The long term fix should be to remove the entity_oids parameter, and force clients to use an id filter as a starting filter in their requests.
For now, we do basically the same thing - just on the server side. If the list of input OIDs is not empty, we convert it to a PropertyFilter and make that filter the starting filter for all the input search parameters.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/32607/