OM-47666 RB:32695 Re-computing the action risk inside AO to incorporate the policy name in it for Reconfigures and Provisions
Background: In M2, we currently just return the BaseType attribute of the commodity that triggered a provision. This helps us in retrieving just the commodityClass. However, we need the Type and baseType to retrieve the class and also the related policy (assuming the baseType corresponded to a segmentationCommodity). This is achieved through returning the commSpec instead (RB:32694). We use this attribute to re-compute the explanation.

Problem: In XL, we need to use the type attribute returned as part of the provision action to display the associated policy.

Solution: 1) The CommSpec sent by M2 attribute is again packaged into the CommodityType in the AO and stored in the ReasonCommodity object in the Explanation and sent to the API component.
2) In the API we retrieve the policyName using the key in the CommType stored in the ReasonCommodity.
3) Recompute the risk string with the policyName.