OM-47666 Returning CommSpec as the reason for Provision so that we can retrieve the associated policy for showing on the action
Background: In M2, we currently just return the BaseType attribute of the commodity that triggered a provision. This helps us in retrieving just the commodityClass. However, we need the Type and baseType to retrieve the class and also the related policy (assuming the baseType corresponded to a segmentationCommodity).

Problem: In XL, we need to use the type attribute returned as part of the provision action to display the associated policy.

Solution: return commSpec as the congestedCommodity that triggers provision instead of just the baseType.